test(mapgen): keep artifact guards package-owned
