+++ docs/projects/standard-recipe-authoring-surface/proof-ledger.md
@@ -158,8 +158,9 @@
 | Stale identity test repair | `bun test mods/mod-swooper-maps/test/config/shipped-map-identity.test.ts` | Passed 1 test / 108 expects. The test now asserts public shipped configs expose semantic Ecology feature authoring, while compiled internal planners preserve the tuned vegetation, wetland, ice, and reef strategy configs. |
 | Source authoring surface guards | `bun test mods/mod-swooper-maps/test/config/standard-authoring-surface-guards.test.ts` | Passed 2 tests / 248 expects. All 17 standard stages use `semantic-public-config`, recursive public object schemas are strict, approved public keys match source, no public raw `{ strategy, config }`, and focus paths do not point at raw envelope keys. |
 | Generated map/SDK boundary guards | `standard-authoring-surface-guards.test.ts` | Generated map entrypoints match canonical config ids, import `../configs/<id>.config.json`, call exactly `canonicalRecipeConfig<StandardRecipeConfig>(mapConfig)` as the SDK `createMap` config expression, record exact recomputed source `configHash`/`envelopeHash`, and do not inline raw placement configs or strategy/config envelopes. |
-| Studio artifact guards | `bun test apps/mapgen-studio/test/config/standardRecipeArtifactGuards.test.ts` | Passed 4 tests / 155 expects. Generated standard schema and UI metadata match source recipe stages, generated default config matches normalized `swooper-earthlike.config.json`, catalog/runtime entries share the same generated artifacts, UI focus paths land on public schema/default paths, and built-in presets validate without raw envelopes. |
-| Shipped config, preset, Studio, and guard suite | `bun test mods/mod-swooper-maps/test/config/maps-schema-valid.test.ts mods/mod-swooper-maps/test/config/presets-schema-valid.test.ts mods/mod-swooper-maps/test/config/studio-presets-schema-valid.test.ts mods/mod-swooper-maps/test/config/shipped-map-identity.test.ts mods/mod-swooper-maps/test/config/standard-authoring-surface-guards.test.ts apps/mapgen-studio/test/config/defaultConfigSchema.test.ts apps/mapgen-studio/test/config/standardRecipeArtifactGuards.test.ts` | Passed 69 tests / 1981 expects after accepted P2 repairs. |
+| Generated recipe artifact guards | `bun run --cwd mods/mod-swooper-maps build:studio-recipes`; `bun test mods/mod-swooper-maps/test/config/standard-recipe-artifact-guards.test.ts` | Passed 1 test / 5 expects. Generated standard schema and UI metadata match source recipe stages, generated default config matches normalized `swooper-earthlike.config.json`, and generated defaults contain no raw envelopes. |
+| Studio artifact guards | `bun run --cwd apps/mapgen-studio test -- standardRecipeArtifactGuards` | Passed 3 tests. Catalog/runtime entries share the same generated artifacts, UI focus paths land on public schema/default paths, and built-in presets validate without raw envelopes. |
+| Shipped config, preset, Studio, and guard suite | `bun test mods/mod-swooper-maps/test/config/maps-schema-valid.test.ts mods/mod-swooper-maps/test/config/presets-schema-valid.test.ts mods/mod-swooper-maps/test/config/studio-presets-schema-valid.test.ts mods/mod-swooper-maps/test/config/shipped-map-identity.test.ts mods/mod-swooper-maps/test/config/standard-authoring-surface-guards.test.ts mods/mod-swooper-maps/test/config/standard-recipe-artifact-guards.test.ts apps/mapgen-studio/test/config/defaultConfigSchema.test.ts apps/mapgen-studio/test/config/standardRecipeArtifactGuards.test.ts` | Focused suite covers shipped configs, generated recipe artifacts, Studio catalog/runtime artifacts, public focus paths, and raw-envelope regressions. |
 | OpenSpec validation | `bun run openspec -- validate studio-sdk-authoring-surface-guards --strict` | Passed. |
 | Package TypeScript check | `bun run --cwd apps/mapgen-studio check` | Passed. |
 | Package build/check | `bun run --cwd packages/mapgen-core build`; `bun run --cwd packages/mapgen-core check` | Passed. |
